+++ PSPDFKit.framework/Versions/A/Headers/PSPDFDocument.h
@@ -233,17 +233,11 @@
 
 /// @name PDF Page Rendering
 
-// PDF rendering options
-extern NSString *kPSPDFIgnoreDisplaySettings; // Always draw pixels with a 1.0 scale.
-extern NSString *kPSPDFPageColor;             // Multiplies a color used to color a page.
-extern NSString *kPSPDFContentOpacity;        // Opacity of the pdf content can be ajusted.
-extern NSString *kPSPDFInvertRendering;       // Inverts the rendering output.
-
 /// Renders the page or a part of it with default display settings into a new image.
 /// @param fullSize		 The size of the page, in pixels, if it was rendered without clipping
 /// @param clippedToRect A rectangle, relative to fullSize, that specifies the area of the page that should be rendered
 /// @param annotations   Annotations that should be rendered with the view
-/// @param options       Dictionary with options that modify the render process.
+/// @param options       Dictionary with options that modify the render process (see PSPDFPageRenderer)
 /// @returns			A new UIImage with the rendered page content
 - (UIImage *)renderImageForPage:(NSUInteger)page withSize:(CGSize)fullSize clippedToRect:(CGRect)clipRect withAnnotations:(NSArray *)annotations options:(NSDictionary *)options;
 
@@ -252,7 +246,8 @@ extern NSString *kPSPDFInvertRendering;       // Inverts the rendering output.
 
 /// Will be called within renderPage:inContex: but before options like inversion are applied.
 /// Note: most of the time you might be better off using custom UIViews and adding them on PSPDFPageView via the delegates.
-- (void)drawOverlayForPage:(NSUInteger)page inContext:(CGContextRef)context withSize:(CGSize)size clippedToRect:(CGRect)clipRect withAnnotations:(NSArray *)annotations options:(NSDictionary *)options;
+
+//- (void)drawOverlayForPage:(NSUInteger)page inContext:(CGContextRef)context withSize:(CGSize)size clippedToRect:(CGRect)clipRect withAnnotations:(NSArray *)annotations options:(NSDictionary *)options;
 
 @end