+++ PSPDFKit.framework/Versions/A/Headers/PSPDFKitGlobal.h
@@ -122,6 +122,7 @@ typedef void(^PSPDFBasicBlock)(void);
 #define PSRectClearCoords(_CGRECT) CGRectMake(0, 0, _CGRECT.size.width, _CGRECT.size.height)
 #define PSIsIpad() ([[UIDevice currentDevice] userInterfaceIdiom] == UIUserInterfaceIdiomPad)
 #define ps_swapf(a,b) { float c = (a); (a) = (b); (b) = c; }
+#define BOXED(val) ({ typeof(val) _tmp_val = (val); [NSValue valueWithBytes:&(_tmp_val) objCType:@encode(typeof(val))]; })
 
 #define PSPDF_KEYPATH(object, property) ((void)(NO && ((void)object.property, NO)), @#property)
 #define PSPDF_KEYPATH_SELF(property) PSPDF_KEYPATH(self, property)