+++ PSPDFKit.framework/Versions/A/Headers/PSPDFTabbedViewController.h
@@ -16,7 +16,7 @@
 @optional
 
 /// Will be called when the documents array changes.
-- (BOOL)tabbedPDFController:(PSPDFTabbedViewController *)tabbedPDFController willChangeDocuments:(NSArray *)newDocuments;
+- (BOOL)tabbedPDFController:(PSPDFTabbedViewController *)tabbedPDFController shouldChangeDocuments:(NSArray *)newDocuments;
 
 /// Will be called after the documents array changed.
 - (void)tabbedPDFController:(PSPDFTabbedViewController *)tabbedPDFController didChangeDocuments:(NSArray *)oldDocuments;